[issue-11] extract warp_pov to own module, add tests
