+++ src/fractals/mod.rs
@@ -4,5 +4,6 @@ pub mod image;
 pub mod job;
 pub mod parser;
 pub mod size;
+pub mod warp_pov;
 
 pub use self::job::Job;
